requirements.txt
django>=3.0,<4.0
pymongo==3.12.3
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.1.0
celery>=5.2.3
beautifulsoup4>=4.10.0
requests==2.24.0
validators==0.18.2
python-dateutil==2.8.2
sorl-thumbnail==12.7.0
